Renamed packages and fixed all imports
